Step_1/requirements.txt
# Default Dependencies
pydantic>=2.0.0  # For data validation
google-generativeai>=0.7.0  # For Gemini API
sentence-transformers>=2.2.0  # For open-source embeddings
lancedb==0.22.0
docling==2.31.0

